test: merge TransitQueueVehicleTest and AbstractTransitVehicleTest into TransitVehicleTest
